Better tests

+++ source/mtree_int8.h
@@ -13,10 +13,10 @@
 typedef struct {
 	int parentDistance;
 	int coveringRadius;
-	int64 data;
+	long long int data;
 } mtree_int8;
 
-#define MTREE_INT8_SIZE (2 * sizeof(int) + sizeof(int64))
+#define MTREE_INT8_SIZE (2 * sizeof(int) + sizeof(long long int))
 #define DatumGetMtreeInt8(x) ((mtree_int8 *) PG_DETOAST_DATUM(x))
 #define PG_GETARG_MTREE_INT8_P(x) DatumGetMtreeInt8(PG_GETARG_DATUM(x))
 #define PG_RETURN_MTREE_INT8_P(x) PG_RETURN_POINTER(x)
